watch: implement the watcher type
This change request adds watches.  Watches enable monitoring a kv node
in consul for changes.  The mechanism consul exposes for doing this is
HTTP long polling.  This means that we have long running connections and
we could potentially miss updates, but it's the only option right now.

Keep an eye on this:

https://github.com/hashicorp/consul/issues/1761