watcher: fix issue where index was not being set
When a call to the consul API results in a 404, we return an error and
don't parse the response headers.  We need the index from the repsonse
headers to prevent pummeling the API with 404s.  The blocking query will
only block when the node doesn't exist if the index is >= X-Consul-Index